requirements.txt
pydantic==1.9.0
thinc==8.0.17
spacy==3.2.0
openai==0.27.0
langchain-google-genai==1.0.7
flask==2.1.3
faiss-cpu==1.7.2
google-generativeai==0.1.0
